Change MAX_CHOICE_NUM to max UINT8 value
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@7122 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdeModulePkg/Universal/PlatformDriverOverride/PlatOverMngr/PlatOverMngr.c
@@ -975,7 +975,7 @@ UpdatePrioritySelectPage (
     EFI_IFR_FLAG_RESET_REQUIRED,
     0,
     EFI_IFR_NUMERIC_SIZE_1,
-    MAX_CHOICE_NUM,
+    (UINT8) MAX_CHOICE_NUM,
     IfrOptionList,
     SelectedDriverImageNum,
     &UpdateData
